+++ CHANGELOG.md
@@ -214,6 +214,9 @@ We change btn colors:
 
 * fix(Button/Link): merge `active-class` / `inactive-class` with app config (#4446)
 
+### From Nuxt List
+* feat(Popover): add `reference` prop
+
 ## [0.7.2](https://github.com/bitrix24/b24ui/compare/v0.7.1...v0.7.2) (2025-07-14)
 
 ### Bug Fixes